Remove the readline shell machinery
IPython now uses the prompt_toolkit based TerminalInteractiveShell
instead, so this removes the readline code we're no longer using.
